+++ utest/namespace/test_keywords.py
@@ -1,3 +1,4 @@
+from builtins import object
 import unittest
 from robotide.namespace.namespace import _Keywords
 from nose.tools import assert_true, assert_false, assert_equals
@@ -8,7 +9,7 @@ class ItemMock(object):
     def __init__(self, name, args, long):
         self.name = name
         self.arguments = args
-        self.longname = long
+        self.longname = int
 
 
 class TestKeywords(unittest.TestCase):
